Add code to update lot usage whenever dump_recursively is called
This commit adds a header file "XrdPfcLot.hh" defining the LotUpdate class, which can be used by "XrdPfcPurge.cc"'s dump_recursively
method to build the appropriate lot update object. This object is then used in LotUpdate's destructor to push lot usage updates to the lot db.

One caveat with the current implementation is that dump_recursively is called before a purge, but not after. As such, lot usage is updated before
any files are purged, but it is not updated after, resulting in the usage reporting to lag by the duration of the interval between purges.